+++ database/migrations/2022_05_18_183501_create_people_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreatePeopleTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
@@ -13,7 +13,8 @@ class CreatePeopleTable extends Migration
      */
     public function up()
     {
-        Schema::create('people', function (Blueprint $table) {
+        if (!Schema::hasTable('people')) {
+            Schema::create('people', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->string('bn_name')->nullable();
@@ -28,6 +29,7 @@ class CreatePeopleTable extends Migration
             $table->bigInteger('blood_group')->nullable();
             $table->timestamps();
         });
+        }
     }
 
     /**
@@ -39,4 +41,4 @@ class CreatePeopleTable extends Migration
     {
         Schema::dropIfExists('people');
     }
-}
+};
