+++ database/migrations/2022_06_14_184231_create_organizations_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateOrganizationsTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
@@ -13,7 +13,8 @@ class CreateOrganizationsTable extends Migration
      */
     public function up()
     {
-        Schema::create('organizations', function (Blueprint $table) {
+        if (!Schema::hasTable('organizations')) {
+            Schema::create('organizations', function (Blueprint $table) {
             $table->id();
             $table->foreignId('institute_id')->constrained('institutes')->onDelete('cascade');
             $table->string('system_id')->unique();
@@ -44,6 +45,7 @@ class CreateOrganizationsTable extends Migration
             $table->boolean('status')->default(true);
             $table->timestamps();
         });
+        }
     }
 
     /**
@@ -55,4 +57,4 @@ class CreateOrganizationsTable extends Migration
     {
         Schema::dropIfExists('organizations');
     }
-}
+};
